+++ CHANGELOG.md
@@ -1,6 +1,56 @@
+# 📅 18/09/2025
+
+## ‼ Releases
+
+### 1.0.0
+
+- 🚀 RELEASED FIRST VERSION
+
+## 🆕 Features
+
+### 0.7.0
+
+- ✅ ADDED: Automatic map build
+
+# 📅 15/09/2025
+
+## 🆕 Features
+
+### 0.6.0
+
+- ✅ ADDED: Loop with background music
+
+# 📅 11/09/2025
+
+## 🆕 Features
+
+### 0.5.0
+
+- ✅ ADDED: Crystal vanish effect
+
+# 📅 08/09/2025
+
+## 🆕 Features
+
+### 0.4.0
+
+- ✅ ADDED: Max score logic
+
+# 📅 21/08/2025
+
+## 🆕 Features
+
+### 0.3.0
+
+- ✅ ADDED: Crystals in the UI and scoring logic
+
 # 📅 07/08/2025
 
-## Features
+## 🆕 Features
+
+### 0.2.0
+
+- ✅ ADDED: Restart game logic
 
 ### 0.1.0